requirements.txt
fastapi
uvicorn
strawberry-graphql
pydantic-settings
sqlalchemy[asyncio]
asyncpg
alembic
bcrypt